% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{add_drugbank_info}
\alias{add_drugbank_info}
\title{add_drugbank_info
  Add passed DrugBank db metadata to passed dvobject}
\usage{
add_drugbank_info(dvobject, parsed_db)
}
\value{
dvobject
}
\description{
add_drugbank_info
  Add passed DrugBank db metadata to passed dvobject
}
\keyword{internal}
