% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{build_metadata}
\alias{build_metadata}
\title{Build Metadata Data Frame}
\usage{
build_metadata(attrs, has_second_db, second_db_name)
}
\arguments{
\item{attrs}{Attributes list}

\item{has_second_db}{Logical indicating if second database exists}

\item{second_db_name}{Name of second database attribute}
}
\value{
Data.frame with database metadata
}
\description{
Build Metadata Data Frame
}
\keyword{internal}
