% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_carrier_node_parser.R
\name{carriers_links}
\alias{carriers_links}
\title{Extracts the drug carriers links element and return data as data frame.}
\usage{
carriers_links(save_table = FALSE, save_csv = FALSE, csv_path = ".",
  override_csv = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed dataframe if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true
in the new parse operation}
}
\value{
drug carriers_links node attributes date frame
}
\description{
\code{parse_drug_groups} returns data frame of drug carriers links elements.
}
\details{
This functions extracts the carriers links element of drug node in \strong{DrugBank}
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{read_drugbank_xml_db}} function like
any other parser function.
If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\dontrun{
# return only the parsed dataframe
carriers_links()

# save in database and return parsed dataframe
carriers_links(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current location
# and return parsed dataframe.
# If the csv exist before read it and return its data.
carriers_links(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not exist in
current
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
carriers_links(ssave_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given location and
# return parsed dataframe.
# If the csv exist before read it and return its data.
carriers_links(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current location and
# return parsed dataframe.
# If the csv exist override it and return it.
carriers_links(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}
}
\seealso{
Other carriers: \code{\link{carriers_actions}},
  \code{\link{carriers_articles}},
  \code{\link{carriers_polypeptide_ext_identity}},
  \code{\link{carriers_polypeptidepeptides_go}},
  \code{\link{carriers_polypeptidepeptides_pfams}},
  \code{\link{carriers_polypeptidepeptides_syn}},
  \code{\link{carriers_polypeptide}},
  \code{\link{carriers_textbooks}}, \code{\link{carriers}}
}
\concept{carriers}
