% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{cett}
\alias{cett}
\title{Run all CETT related parsers}
\usage{
cett()
}
\value{
a list of all drugs parsed tibbles
}
\description{
Run all parsers that retrieve carriers, enzymes, targets and transporters
 related information
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other collective_parsers: 
\code{\link{drugs}()},
\code{\link{run_all_parsers}()}
}
\concept{collective_parsers}
