% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_actions_parsers.R
\name{cett_actions_doc}
\alias{cett_actions_doc}
\alias{carriers_actions}
\alias{enzymes_actions}
\alias{targets_actions}
\alias{transporters_actions}
\title{Carriers/ Enzymes/ Targets/ Transporters Actions parsers}
\usage{
carriers_actions()

enzymes_actions()

targets_actions()

transporters_actions()
}
\value{
a tibble with 2 variables:
\describe{
  \item{action}{describe related action}
  \item{\emph{parent_id}}{carrier/ target/ enzyme/ transporter id}
}
}
\description{
Collection of related actions
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other cett: 
\code{\link{cett_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_doc}},
\code{\link{cett_poly_pfms_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
