% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_others_parsers.R
\name{cett_go_doc}
\alias{cett_go_doc}
\alias{carriers_polypeptides_go}
\alias{enzymes_polypeptides_go}
\alias{targets_polypeptides_go}
\alias{transporters_polypeptides_go}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide GO Classifier
parsers}
\usage{
carriers_polypeptides_go()

enzymes_polypeptides_go()

targets_polypeptides_go()

transporters_polypeptides_go()
}
\value{
a tibble with 3 variables:
\describe{
  \item{category}{}
  \item{description}{}
  \item{parent_key}{polypeptide id}
}
}
\description{
Extract descriptions of identified polypeptide go classifier for targets,
 enzymes, carriers, or transporters.
}
\keyword{internal}
