% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{cett_nodes_options}
\alias{cett_nodes_options}
\title{returns carriers, enzymes,targets and transporters node valid options.}
\usage{
cett_nodes_options()
}
\value{
list of CETT valid options
}
\description{
returns carriers, enzymes,targets and transporters node valid options.
}
\seealso{
Other parsers: 
\code{\link{drug_node_options}()},
\code{\link{parseDrugBank}()},
\code{\link{parseHPO}()},
\code{\link{parseOnSIDES}()},
\code{\link{parseTWOSIDES}()},
\code{\link{references_node_options}()}
}
\concept{parsers}
