% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_general_infromation_parsers.R
\name{cett_poly_doc}
\alias{cett_poly_doc}
\alias{carriers_polypeptides}
\alias{enzymes_polypeptides}
\alias{targets_polypeptides}
\alias{transporters_polypeptides}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide parsers}
\usage{
carriers_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

enzymes_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

targets_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

transporters_polypeptides(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed tibble if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the
new parse operation}
}
\value{
a tibble with 20 variables:
\describe{
  \item{id}{\href{https://www.uniprot.org/}{Universal Protein Resource
  (UniProt) identifier}}
  \item{source}{Specifies whether the identified polypeptide ID is
  associated with any of the following UniProt knowledge bases:
  Swiss-Prot, which is manually annotated and reviewed, or TrEMBL,
  which is automatically annotated and not reviewed.}
  \item{name}{}
  \item{general_function}{General summary of the physiological function of
  the polypeptide}
  \item{specific_function}{A more specific description of the polypeptide’s
   physiological function within the cell.}
  \item{gene_name}{The short name commonly associated with the associated
   gene. Eg. PTGS1.}
  \item{locus}{The specific chromosomal location or position of the gene’s
   sequence on a chromosome.}
  \item{cellular_location}{The cellular location of the polypeptide.}
  \item{transmembrane_regions}{Areas of the polypeptide sequence that span
   a biological membrane.}
  \item{signal_regions}{Location of any signal peptides within the
   polypeptide sequence.}
  \item{theoretical_pi}{Theoretical isoelectric point.}
  \item{molecular_weight}{The molecular weight of the polypeptide.}
  \item{chromosome_location}{The chromosomal location of the polypeptide
  gene}
  \item{organism}{The organism in which this polypeptide functions.}
  \item{organism_ncbi_taxonomy_id}{}
  \item{amino_acid_sequence}{The amino acid sequence of the polypeptide}
  \item{amino_acid_format}{}
  \item{gene_sequence}{The sequence of the associated gene.}
  \item{gene_format}{}
  \item{parent_key}{carrier/ target/ enzyme/ transporter id}
}
}
\description{
Extract descriptions of identified polypeptide targets, enzymes, carriers,
 or transporters.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

# save parsed tibble as csv if it does not exist in current location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE)

# save parsed tibble as csv if it does not exist in given location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in current location and
# return parsed tibble.
# if the csv exist override it and return it.
run_all_parsers(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_pfms_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
