% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_others_parsers.R
\name{cett_poly_pfms_doc}
\alias{cett_poly_pfms_doc}
\alias{carriers_polypeptides_pfams}
\alias{enzymes_polypeptides_pfams}
\alias{targets_polypeptides_pfams}
\alias{transporters_polypeptides_pfams}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide PFAMS parsers}
\usage{
carriers_polypeptides_pfams()

enzymes_polypeptides_pfams()

targets_polypeptides_pfams()

transporters_polypeptides_pfams()
}
\value{
a tibble with 3 variables:
\describe{
  \item{name}{The sequence of the associated gene.}
  \item{identifier}{}
  \item{parent_key}{polypeptide id}
}
}
\description{
Extract descriptions of identified polypeptide PFAMS targets, enzymes,
 carriers, or transporters.
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
