% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_tables_saver.R
\name{close_db}
\alias{close_db}
\title{Close open drug bank sql database}
\usage{
close_db()
}
\description{
\code{close_db} closes connection to pre-given database.
}
\details{
This function closes connection to pre-given databas.
}
\examples{
close_db()
}
