% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{count_top_level_lists}
\alias{count_top_level_lists}
\title{Count Top Level Lists}
\usage{
count_top_level_lists(obj)
}
\arguments{
\item{obj}{Object to analyze}
}
\value{
Integer count of top-level lists (excluding data.frames)
}
\description{
Count Top Level Lists
}
\keyword{internal}
