% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_affected_organisms}
\alias{drug_affected_organisms}
\title{Drug Affected Organism parser}
\usage{
drug_affected_organisms()
}
\value{
a tibble with 2 variables:
\describe{
 \item{affected-organism}{affected-organism name}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Organisms in which the drug may display activity; activity may depend on
local susceptibility patterns and resistance.
}
\keyword{internal}
