% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_ahfs_codes}
\alias{drug_ahfs_codes}
\title{Drug ahfs-codes parser}
\usage{
drug_ahfs_codes()
}
\value{
a tibble with the following variables:
\describe{
 \item{ahfs-code}{}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
The American Hospital Formulary Service (AHFS) identifier for this drug.
}
\keyword{internal}
