% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_atc_codes_node_parser.R
\name{drug_atc_codes}
\alias{drug_atc_codes}
\title{Drug ATC Codes element parser}
\usage{
drug_atc_codes()
}
\value{
a tibble with 10 variables
}
\description{
The Anatomical Therapeutic Classification (ATC) code for the drug assigned
by the World Health Organization Anatomical Chemical Classification System.
}
\details{
Each `atc-code`` row has one or more level. The atc-code and level>
have a code  the code assigned by the World Health Organization Anatomical
Therapeutic Chemical Classification system.
}
\keyword{internal}
