% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_calc_prop}
\alias{drug_calc_prop}
\title{Drug Calculated Properties parser}
\usage{
drug_calc_prop()
}
\value{
a tibble with 4 variables:
\describe{
 \item{kind}{Name of the property.}
 \item{value}{Predicted physicochemical properties; obtained by the use of
 prediction software such as ALGOPS and ChemAxon.}
 \item{source}{Name of the software used to calculate this property,
 either ChemAxon or ALOGPS.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Drug properties that have been predicted by ChemAxon or ALOGPS based on the
imputed chemical structure. Associated links below will redirect to
descriptions of the specific term.
}
\keyword{internal}
