% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_categories}
\alias{drug_categories}
\title{Drug Categories parser}
\usage{
drug_categories()
}
\value{
a tibble with 2 variables:
\describe{
 \item{category}{category name}
 \item{mesh-id}{The Medical Subjects Headings (MeSH) identifier for the
 category.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
General categorizations of the drug.
}
\keyword{internal}
