% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_classfication_node_parser.R
\name{drug_classification}
\alias{drug_classification}
\title{Drug Classification parser}
\usage{
drug_classification()
}
\value{
a tibble with 9 variables:
\describe{
  \item{description}{}
  \item{direct-parent}{}
  \item{kingdom}{}
  \item{superclass}{}
  \item{class}{}
  \item{subclass}{}
  \item{alternative-parent}{One or more alternative parents}
  \item{substituent}{One or more substituents}
  \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A description of the hierarchical chemical classification of the drug;
imported from ClassyFire.
}
\keyword{internal}
