% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_dosages}
\alias{drug_dosages}
\title{Drug Dosages parser}
\usage{
drug_dosages()
}
\value{
a tibble with the following variables:
\describe{
 \item{form}{The pharmaceutical formulation by which the drug is introduced
 into the body}
 \item{route}{The path by which the drug or product is taken into the body.}
 \item{strength}{The amount of active drug ingredient provided in the dosage}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A list of the commercially available dosages of the drug.
}
\keyword{internal}
