% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{drug_element}
\alias{drug_element}
\title{extracts the given drug elements and return data as list of tibbles.}
\usage{
drug_element(elements_options = c("all"))
}
\arguments{
\item{elements_options}{list,  options of elements to be parsed. default is
"all"}
}
\value{
list of selected drug elements tibbles
}
\description{
\code{drug_element} returns list of tibbles of <- selected
elements.
}
\details{
drug_element_options can be called to know the valid options for
this method
}
\examples{
\dontrun{
# return only the parsed tibble
drug_element()
drug_element(c("drug_ahfs_codes", "drug_carriers"))
}
}
\seealso{
Other common: 
\code{\link{drug_element_options}()},
\code{\link{drug_node_options}()},
\code{\link{references_node_options}()},
\code{\link{run_all_parsers}()}
}
\concept{common}
\keyword{internal}
