% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{drug_element_options}
\alias{drug_element_options}
\title{returns \code{drug_element} valid options.}
\usage{
drug_element_options()
}
\value{
list of \code{drug_element} valid options
}
\description{
returns \code{drug_element} valid options.
}
\examples{
\dontrun{
drug_element_options()
}
}
\seealso{
Other common: 
\code{\link{drug_element}()},
\code{\link{run_all_parsers}()}
}
\concept{common}
