% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_external_links}
\alias{drug_external_links}
\title{Drug External Links parser}
\usage{
drug_external_links()
}
\value{
a tibble with the following variables:
\describe{
 \item{resource}{Name of the source website.}
 \item{identifier}{Identifier for this drug in the given resource}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Links to other websites or databases providing information about this drug.
}
\keyword{internal}
