% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_food_interactions}
\alias{drug_food_interactions}
\title{Drug Groups parser}
\usage{
drug_food_interactions()
}
\value{
a tibble with the following variables:
\describe{
 \item{food-interaction}{}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Food that may interact with this drug.
}
\keyword{internal}
