% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_groups}
\alias{drug_groups}
\title{Drug Groups parser}
\usage{
drug_groups()
}
\value{
a tibble with 2 variables:
\describe{
 \item{group}{}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Groups that this drug belongs to. May include any of: approved, vet_approved,
 nutraceutical, illicit, withdrawn, investigational, and experimental.
}
\keyword{internal}
