% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_interactions}
\alias{drug_interactions}
\title{Drug Interactions parser}
\usage{
drug_interactions()
}
\value{
a tibble with the following variables:
\describe{
 \item{drugbank-id	}{DrugBank ID of the interacting drug.}
 \item{name}{Name of the interacting drug.}
 \item{description}{Textual description of the physiological consequences
 of the drug interaction}
 \item{\emph{drugbank_id}}{parent drugbank id}
}
}
\description{
Drug-drug interactions detailing drugs that, when administered concomitantly
with the drug of interest, will affect its activity or result in adverse
effects. These interactions may be synergistic or antagonistic depending on
the physiological effects and mechanism of action of each drug.
}
\keyword{internal}
