% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_intern_brand}
\alias{drug_intern_brand}
\title{Drug International Brands parser}
\usage{
drug_intern_brand()
}
\value{
a tibble with 4 variables:
\describe{
 \item{brand}{The proprietary, well-known name for given to this drug by a
 manufacturer.}
 \item{company}{The company or manufacturer that uses this name.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
The proprietary names used by the manufacturers for commercially available
forms of the drug, focusing on brand names for products that are available
in countries other than Canada and the Unites States.
}
\keyword{internal}
