% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_manufacturer_node_parser.R
\name{drug_manufacturers}
\alias{drug_manufacturers}
\title{Drug Manufacturers parser}
\usage{
drug_manufacturers()
}
\value{
a tibble with the following variables:
\describe{
 \item{generic}{A list of companies that are manufacturing the generic
  form of the drug.}
 \item{url}{A link to the companies that are manufacturing the drug.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A list of companies that are manufacturing the commercially available forms
of this drug that are available in Canada and the Unites States.
}
\keyword{internal}
