% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_mixtures}
\alias{drug_mixtures}
\title{Drug Mixtures parser}
\usage{
drug_mixtures()
}
\value{
a tibble with 4 variables:
\describe{
 \item{name}{The proprietary name provided by the manufacturer for this
 combination product.}
 \item{ingredients}{A list of ingredients, separated by addition symbols}
 \item{supplemental-ingredients}{List of additional active ingredients which
  are not clinically relevant to the main indication of the product,
  separated by addition symbols.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
All commercially available products in which this drug is available in
combination with other drug molecules
}
\keyword{internal}
