% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_packagers}
\alias{drug_packagers}
\title{Drug Packagers parser}
\usage{
drug_packagers()
}
\value{
a tibble with 2 variables:
\describe{
 \item{name}{}
 \item{url}{A link to any companies that are packaging the drug for
 re-distribution.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A list of companies that are packaging the drug for re-distribution.
}
\keyword{internal}
