% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_pathway_node_parsers.R
\name{drug_pathway_drugs}
\alias{drug_pathway_drugs}
\title{Drug Pathway Drugs parser}
\usage{
drug_pathway_drugs()
}
\value{
a tibble with pathway drugs properties
}
\description{
Drugs involved in this pathway.
}
\keyword{internal}
