% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_pathway_node_parsers.R
\name{drug_pathway_enzyme}
\alias{drug_pathway_enzyme}
\title{Drug Pathway Enzymes parser}
\usage{
drug_pathway_enzyme()
}
\value{
a tibble with pathway properties
}
\description{
Enzymes involved in this pathway.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other pathway: 
\code{\link{drug_pathway_drugs}()},
\code{\link{drug_pathway}()}
}
\concept{pathway}
