% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_pdb_entries}
\alias{drug_pdb_entries}
\title{Drug pdb-entries parser}
\usage{
drug_pdb_entries()
}
\value{
a tibble with the following variables:
\describe{
 \item{pdb-entry}{}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Protein Data Bank (PDB) identifiers for this drug.
}
\keyword{internal}
