% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_price_node_parser.R
\name{drug_prices}
\alias{drug_prices}
\title{Drug Prices Parsers}
\usage{
drug_prices()
}
\value{
a tibble with 5 variables:
\describe{
  \item{description}{}
  \item{cost}{Drug price per unit}
  \item{currency}{Currency of price, example: US.}
  \item{unit}{}
  \item{parent_id}{drugbank id}
}
}
\description{
Unit drug prices
}
\keyword{internal}
