% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_reaction_node_parser.R
\name{drug_reactions_enzymes}
\alias{drug_reactions_enzymes}
\title{Drug Reactions Enzymes Parsers}
\usage{
drug_reactions_enzymes()
}
\value{
a tibble with 3 variables:
\describe{
  \item{name}{}
  \item{uniprot-id}{}
  \item{parent_id}{drugbank id}
}
}
\description{
Enzymes involved in metabolizing this drug
}
\keyword{internal}
