% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_snp_adverse_reactions}
\alias{drug_snp_adverse_reactions}
\title{Drug SNP Adverse Drug Reactions parser}
\usage{
drug_snp_adverse_reactions()
}
\value{
a tibble with the following variables:
\describe{
 \item{protein-name}{Proteins involved in this SNP.}
 \item{gene-symbol}{Genes involved in this SNP.}
 \item{uniprot-id}{Universal Protein Resource (UniProt) identifiers for
 proteins involved in this pathway.}
 \item{rs-id}{The SNP Database identifier for this single nucleotide
  polymorphism.}
 \item{allele}{The alleles associated with the identified SNP.}
 \item{adverse-reaction}{}
 \item{description}{}
 \item{pubmed-id	}{Reference to PubMed article.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
The adverse drug reactions that may occur as a result of the listed single
nucleotide polymorphisms (SNPs)
}
\keyword{internal}
