% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_snp_effects}
\alias{drug_snp_effects}
\title{Drug SNP Effects parser}
\usage{
drug_snp_effects()
}
\value{
a tibble with the following variables:
\describe{
 \item{protein-name}{Proteins involved in this SNP.}
 \item{gene-symbol}{Genes involved in this SNP.}
 \item{uniprot-id}{Universal Protein Resource (UniProt) identifiers for
 proteins involved in this pathway.}
 \item{rs-id}{	The SNP Database identifier for this single nucleotide
 polymorphism.}
 \item{allele}{The alleles associated with the identified SNP.}
 \item{defining-change}{}
 \item{description}{A written description of the SNP effects.}
 \item{pubmed-id	}{Reference to PubMed article.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A list of single nucleotide polymorphisms (SNPs) relevant to drug activity or
 metabolism, and the effects these may have on pharmacological activity.
 SNP effects in the patient may require close monitoring, an increase or
}
\keyword{internal}
