% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_synonym_node_parser.R
\name{drug_syn}
\alias{drug_syn}
\title{Drug Synonyms parser}
\usage{
drug_syn()
}
\value{
a tibble with 3 variables:
\describe{
 \item{language}{Names of the drug in languages other than English.}
 \item{coder}{Organisation or source providing the synonym. For example,
  INN indicates the synonym is an International Nonproprietary Name,
  while IUPAC indicates the synonym is the nomenclature designated by the
  International Union of Pure and Applied Chemistry.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Other names or identifiers that are associated with this drug.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()}
}
\concept{drugs}
