% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_enzyme_node_parser.R
\name{enzymes_polypeptide_pfams}
\alias{enzymes_polypeptide_pfams}
\title{Extracts the drug enzymes polypeptides pfams element and return
 data as tibble.}
\usage{
enzymes_polypeptide_pfams(save_table = FALSE, save_csv = FALSE,
  csv_path = ".", override_csv = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed tibble if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the
new parse operation}
}
\value{
drug groups node attributes date frame
}
\description{
\code{enzymes_polypeptide_pfams} returns tibble of drug
enzymes polypeptides pfams elements.
}
\details{
This functions extracts the enzymes polypeptides pfams element of drug
 node in \strong{DrugBank}
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned tibble in the database.
It must be called after \code{\link{read_drugbank_xml_db}} function like
any other parser function.
If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\dontrun{
# return only the parsed tibble
enzymes_polypeptide_pfams()

# save in database and return parsed tibble
enzymes_polypeptide_pfams(save_table = TRUE)

# save parsed tibble as csv if it does not exist in current
# location and return parsed tibble.
# If the csv exist before read it and return its data.
enzymes_polypeptide_pfams(save_csv = TRUE)

# save in database, save parsed tibble as csv if it does not exist
# in current location and return parsed tibble.
# If the csv exist before read it and return its data.
enzymes_polypeptide_pfams(save_table = TRUE, save_csv = TRUE)

# save parsed tibble as csv if it does not exist in given
# location and return parsed tibble.
# If the csv exist before read it and return its data.
enzymes_polypeptide_pfams(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in
# current location and return parsed tibble.
# If the csv exist override it and return it.
enzymes_polypeptide_pfams(
  save_csv = TRUE, csv_path = TRUE,
  override = TRUE
)
}
}
\seealso{
Other enzymes: \code{\link{enzymes_actions}},
  \code{\link{enzymes_articles}},
  \code{\link{enzymes_links}},
  \code{\link{enzymes_polypeptide_ext_ident}},
  \code{\link{enzymes_polypeptide_go}},
  \code{\link{enzymes_polypeptide_syn}},
  \code{\link{enzymes_polypeptide}},
  \code{\link{enzymes_textbooks}}, \code{\link{enzymes}}
}
\concept{enzymes}
