% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{find_dataframes_recursive}
\alias{find_dataframes_recursive}
\title{Find All Data Frames Recursively}
\usage{
find_dataframes_recursive(x, prefix = "")
}
\arguments{
\item{x}{Object to search}

\item{prefix}{Current path prefix}
}
\value{
Data.frame with paths, dimensions, and sizes
}
\description{
Find All Data Frames Recursively
}
\keyword{internal}
