% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{find_second_database}
\alias{find_second_database}
\title{Find Second Database in Attributes}
\usage{
find_second_database(attrs)
}
\arguments{
\item{attrs}{Attributes list}
}
\value{
List with has_second_db (logical) and second_db_name (character)
}
\description{
Find Second Database in Attributes
}
\keyword{internal}
