% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_metadata.R
\name{get_drugbank_metadata}
\alias{get_drugbank_metadata}
\title{Return uploaded drugbank database metadata}
\usage{
get_drugbank_metadata()
}
\value{
drugbank metadata
}
\description{
\code{get_drugbank_metadata} returns uploaded drugbank database version and
exported date.
}
\examples{
\donttest{
get_drugbank_metadata()
}
}
