% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_metadata.R
\name{get_drugbank_version}
\alias{get_drugbank_version}
\title{Return uploaded drugbank database version}
\usage{
get_drugbank_version()
}
\value{
drugbank version
}
\description{
\code{get_drugbank_version} returns uploaded drugbank database version.
}
\examples{
\dontrun{
get_drugbank_version()
}
}
