% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_parsers.R
\name{get_genes_for_phenotype}
\alias{get_genes_for_phenotype}
\title{Get genes for a specific phenotype}
\usage{
get_genes_for_phenotype(hpo_db, phenotype_name, exact = FALSE)
}
\arguments{
\item{hpo_db}{An HPO dvobject.}

\item{phenotype_name}{Character. Exact or partial phenotype name.}

\item{exact}{Logical. If TRUE, match exactly. If FALSE, partial match.}
}
\value{
Tibble of genes associated with the phenotype.
}
\description{
Get genes for a specific phenotype
}
