% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{init_dvobject}
\alias{init_dvobject}
\title{init_dvobject
  initialize dvobject}
\usage{
init_dvobject()
}
\value{
dvobject
}
\description{
init_dvobject
  initialize dvobject
}
\keyword{internal}
