% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{links}
\alias{links}
\alias{drugs_links}
\alias{carriers_links}
\alias{enzymes_links}
\alias{targets_links}
\alias{transporters_links}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters links element parser}
\usage{
drugs_links(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

carriers_links(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

enzymes_links(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

targets_links(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)

transporters_links(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE
)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed tibble if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the
new parse operation}
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Name of the source website}
  \item{title}{Identifier for this drug in the given resource}
  \item{url}{The url of the website}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}
}
\description{
Return a list of websites that were used as references for
Drugs/ Carriers/ Enzymes/ Targets/ Transporters
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

# save parsed tibble as csv if it does not exist in current location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE)

# save parsed tibble as csv if it does not exist in given location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
run_all_parsers(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in current location and
# return parsed tibble.
# if the csv exist override it and return it.
run_all_parsers(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{attachments}},
\code{\link{books}},
\code{\link{references}()}
}
\concept{references}
