% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_common_utilities.R
\name{open_mdb}
\alias{open_mdb}
\title{Establish connection to given data base}
\usage{
open_mdb(username = "root", password = "root", server, output_database,
  host = "localhost", port = 3306)
}
\arguments{
\item{server}{string, indicated the db server name.}

\item{output_database}{string, the database name to be used,
it has to be created before using it}

\item{driver}{odbc object to define database driver.}

\item{trusted_connection}{boolean, is the connection secure}
}
\value{
sets the open connection in memory to be used by other functions
}
\description{
\code{open_mdb} opens connection to given database.
}
\details{
This function establishes connection to given database
to store, \emph{optionally}, the parsed drug bank elements.
}
\examples{
\donttest{
open_mdb(driver = "SQL Server",
server = "MOHAMMED\\\\\\\\SQL2016", output_database = "drugbank2")
}
}
