% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{parse_cett_node}
\alias{parse_cett_node}
\title{Run all CETT related parsers}
\usage{
parse_cett_node(cett_options)
}
\value{
a list of all drugs parsed tibbles
}
\description{
Run all parsers that retrieve carriers, enzymes, targets and transporters
 related information
}
\keyword{internal}
