% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R, R/drug_parsers.R
\name{parse_drug}
\alias{parse_drug}
\title{Extracts the all drug elements and return data as list of dataframes.}
\usage{
parse_drug(save_table = FALSE)

parse_drug(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_table}{boolean, save table in database if true.}
}
\value{
drug main node attributs data frame

drug main node attributs data frame
}
\description{
\code{parse_drug_all} returns list of dataframes of drugs elements.

\code{parse_drug} returns data frame of drugs main elements.
}
\details{
This functions extracts all element of drug nodes in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this funtion.

This functions extracts the main element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this funtion.
}
\examples{
parse_drug_all()
parse_drug_all(TRUE)
parse_drug_all(save_table = FALSE)
parse_drug()
parse_drug(TRUE)
parse_drug(save_table = FALSE)
}
