% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_atc_codes_node_parser.R
\name{parse_drug_atc_codes}
\alias{parse_drug_atc_codes}
\title{Extracts the drug atc codes element and return data as data frame.}
\usage{
parse_drug_atc_codes(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug atc_codes node attributs date frame
}
\description{
\code{parse_drug_atc_codes} returns data frame of drug atc codes elements.
}
\details{
This functions extracts the atc codes element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this funtion.
}
\examples{
parse_drug_atc_codes()
parse_drug_atc_codes(TRUE)
parse_drug_atc_codes(save_table = FALSE)
}
