% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_carrier_node_parser.R
\name{parse_drug_carriers_links}
\alias{parse_drug_carriers_links}
\title{Extracts the drug carriers links element and return data as data frame.}
\usage{
parse_drug_carriers_links(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug carriers_links node attributs date frame
}
\description{
\code{parse_drug_groups} returns data frame of drug carriers links elements.
}
\details{
This functions extracts the carriers links element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this funtion.
}
\examples{
\dontrun{
parse_drug_carriers_links()
parse_drug_carriers_links(TRUE)
parse_drug_carriers_links(save_table = FALSE)
}
}
