% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{parse_drug_experimental_properties}
\alias{parse_drug_experimental_properties}
\title{Extracts the drug experimental properties element and return data as data frame.}
\usage{
parse_drug_experimental_properties(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug experimental properties node attributs date frame
}
\description{
\code{parse_drug_experimental_properties} returns data frame of drug experimental
 properties elements.
}
\details{
This functions extracts the experimental properties element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this funtion.
}
\examples{
parse_drug_experimental_properties()
parse_drug_experimental_properties(TRUE)
parse_drug_experimental_properties(save_table = FALSE)
}
