% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{parse_drug_nodes}
\alias{parse_drug_nodes}
\title{Run all drug  related parsers}
\usage{
parse_drug_nodes(drug_options)
}
\arguments{
\item{drug_options}{- list of options to parse}
}
\value{
a list of all drugs parsed tibbles
}
\description{
Run all parsers that retrieve drugs related information
}
\keyword{internal}
