% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{parse_drug_snp_effects}
\alias{parse_drug_snp_effects}
\title{Extracts the drug snp effects element and return data as data frame.}
\usage{
parse_drug_snp_effects(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug snp effects node attributes date frame
}
\description{
\code{parse_drug_snp_effects} returns data frame of snp effects groups elements.
}
\details{
This functions extracts the snp effects element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\dontrun{
parse_drug_snp_effects()
parse_drug_snp_effects(TRUE)
parse_drug_snp_effects(save_table = FALSE)
}
}
