% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_targets_node_parser.R
\name{parse_drug_targets_polypeptides_external_identifiers}
\alias{parse_drug_targets_polypeptides_external_identifiers}
\title{Extracts the drug targets polypeptides external identifiers
 element and return data as data frame.}
\usage{
parse_drug_targets_polypeptides_external_identifiers(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug targets polypeptides external identifiers node attributes date frame
}
\description{
\code{parse_drug_targets_polypeptides_external_identifiers}
 returns data frame of drug targets polypeptides external identifiers elements.
}
\details{
This functions extracts the targets polypeptides external identifiers
 element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\donttest{
parse_drug_targets_polypeptides_external_identifiers()
parse_drug_targets_polypeptides_external_identifiers(TRUE)
parse_drug_targets_polypeptides_external_identifiers(save_table = FALSE)
}
}
