% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_transporter_node_parser.R
\name{parse_drug_transporters_links}
\alias{parse_drug_transporters_links}
\title{Extracts the drug transporters links element and return data as data frame.}
\usage{
parse_drug_transporters_links(save_table = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}
}
\value{
drug transporters links node attributes date frame
}
\description{
\code{parse_drug_transporters_links} returns data frame of drug transporters_ inks elements.
}
\details{
This functions extracts the transporters links element of drug node in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\donttest{
parse_drug_transporters_links()
parse_drug_transporters_links(TRUE)
parse_drug_transporters_links(save_table = FALSE)
}
}
