% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_enzyme_node_parser.R
\name{parse_enzy_poly}
\alias{parse_enzy_poly}
\title{Extracts the drug enzymes polypeptides element and return data as data frame.}
\usage{
parse_enzy_poly(save_table = FALSE, save_csv = FALSE, csv_path = ".",
  override_csv = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed dataframe if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true
in the new parse operation}
}
\value{
drug enzymes polypeptides node attributes date frame
}
\description{
\code{parse_enzy_poly} returns data frame of drug enzymes
polypeptides elements.
}
\details{
This functions extracts the enzymes polypeptides element of drug node in
drugbank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\donttest{
# return only the parsed dataframe
parse_enzy_poly()

# save in database and return parsed dataframe
parse_enzy_poly(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_enzy_poly(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not
# exist in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_enzy_poly(ssave_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_enzy_poly(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist override it and return it.
parse_enzy_poly(
  save_csv = TRUE, csv_path = TRUE,
  override = TRUE
)
}
}
