% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{parse_references_node}
\alias{parse_references_node}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters references element parser}
\usage{
parse_references_node(references_options)
}
\description{
Return a list of all references for drugs, carriers, enzymes, targets or
transporters
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{attachments}},
\code{\link{books}},
\code{\link{links}}
}
\concept{references}
