% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{references}
\alias{references}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters references element parser}
\usage{
references()
}
\description{
Return a list of all references for drugs, carriers, enzymes, targets or
transporters
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{attachments}},
\code{\link{books}},
\code{\link{links}}
}
\concept{references}
