% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{references_node_options}
\alias{references_node_options}
\title{returns references node valid options.}
\usage{
references_node_options()
}
\value{
list of references valid options
}
\description{
returns references node valid options.
}
\seealso{
Other parsers: 
\code{\link{cett_nodes_options}()},
\code{\link{drug_node_options}()},
\code{\link{parseDrugBank}()},
\code{\link{parseHPO}()},
\code{\link{parseOnSIDES}()},
\code{\link{parseTWOSIDES}()}
}
\concept{parsers}
