% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{run_all_parsers}
\alias{run_all_parsers}
\title{extracts the all drug elements and return data as list of tibbles.}
\usage{
run_all_parsers()
}
\value{
all drug elements tibbles
}
\description{
this functions extracts all element of drug nodes in \strong{DrugBank}
xml database.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other common: 
\code{\link{drug_element_options}()},
\code{\link{drug_element}()}

Other collective_parsers: 
\code{\link{cett}()},
\code{\link{drugs}()}
}
\concept{collective_parsers}
\concept{common}
