% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_targets_node_parser.R
\name{targets_polypeptide}
\alias{targets_polypeptide}
\title{Extracts the drug targ polypeptides element and return data as data frame.}
\usage{
targets_polypeptide(save_table = FALSE, save_csv = FALSE,
  csv_path = ".", override_csv = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed dataframe if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the
new parse operation}
}
\value{
drug targ polypeptides node attributes date frame
}
\description{
\code{targets_polypeptide} returns data frame of drug targ
 polypeptides elements.
}
\details{
This functions extracts the targ polypeptides element of drug node
in \strong{DrugBank}
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{read_drugbank_xml_db}} function like
any other parser function.
If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\dontrun{
# return only the parsed dataframe
targets_polypeptide()

# save in database and return parsed dataframe
targets_polypeptide(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
targets_polypeptide(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not
# exist in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
targets_polypeptide(save_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given
# location and return parsed dataframe.
# If the csv exist before read it and return its data.
targets_polypeptide(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist override it and return it.
targets_polypeptide(
  save_csv = TRUE, csv_path = TRUE,
  override = TRUE
)
}
}
\seealso{
Other targets: \code{\link{targets_actions}},
  \code{\link{targets_articles}},
  \code{\link{targets_links}},
  \code{\link{targets_polypeptide_ext_ident}},
  \code{\link{targets_polypeptide_go}},
  \code{\link{targets_polypeptide_pfams}},
  \code{\link{targets_polypeptide_syn}},
  \code{\link{targets_textbooks}}, \code{\link{targets}}
}
\concept{targets}
