% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{add_database_info}
\alias{add_database_info}
\title{add_database_info
  Assign passed databases db metadata to passed dvobject}
\usage{
add_database_info(
  dvobject,
  db_type = "DrugBank",
  db_version = NULL,
  db_exported_date = NULL
)
}
\arguments{
\item{dvobject}{dvobject to assign metadata to it}

\item{db_type}{database type (defualt="DrugBank")}

\item{db_version}{database version as string}

\item{db_exported_date}{database official export date}
}
\value{
dvobject
}
\description{
add_database_info
  Assign passed databases db metadata to passed dvobject
}
