% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{articles}
\alias{articles}
\alias{drugs_articles}
\alias{carriers_articles}
\alias{enzymes_articles}
\alias{targets_articles}
\alias{transporters_articles}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters articles element parser}
\usage{
drugs_articles()

carriers_articles()

enzymes_articles()

targets_articles()

transporters_articles()
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article,
  attachments).}
  \item{pubmed-id}{The PubMed identifier for the article.}
  \item{citation}{Article citation in a standard format.}
  \item{\emph{parent_id}}{drug/carrier/target/enzyme/transporter id}
}
}
\description{
Return a list of articles that were used as references for drugs carriers
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other references: 
\code{\link{attachments}},
\code{\link{books}},
\code{\link{links}},
\code{\link{parse_references_node}()}
}
\concept{references}
