% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{attachments}
\alias{attachments}
\alias{drugs_attachments}
\alias{carriers_attachments}
\alias{enzymes_attachments}
\alias{targets_attachments}
\alias{transporters_attachments}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters attachments element parser}
\usage{
drugs_attachments()

carriers_attachments()

enzymes_attachments()

targets_attachments()

transporters_attachments()
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article,
  attachments).}
  \item{title}{The title of the attachment.}
  \item{url}{The url to download the attachment from.}
  \item{\emph{parent_id}}{drug/carrier/target/enzyme/transporter id}
}
}
\description{
Return a list of attachment that were used as references for drugs carriers
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{books}},
\code{\link{links}},
\code{\link{references}()}
}
\concept{references}
