% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{books}
\alias{books}
\alias{drugs_textbooks}
\alias{carriers_textbooks}
\alias{enzymes_textbooks}
\alias{targets_textbooks}
\alias{transporters_textbooks}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters books element parser}
\usage{
drugs_textbooks()

carriers_textbooks()

enzymes_textbooks()

targets_textbooks()

transporters_textbooks()
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article,
  attachments).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}
}
\description{
Return a list of text books that were used as references for drugs, carriers,
 enzymes, targets or transporters
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{attachments}},
\code{\link{links}},
\code{\link{parse_references_node}()}
}
\concept{references}
