% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{books}
\alias{books}
\alias{drugs_textbooks}
\alias{carriers_textbooks}
\alias{enzymes_textbooks}
\alias{targets_textbooks}
\alias{transporters_textbooks}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters books element parser}
\usage{
drugs_textbooks(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

carriers_textbooks(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

enzymes_textbooks(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

targets_textbooks(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)

transporters_textbooks(
  save_table = FALSE,
  save_csv = FALSE,
  csv_path = ".",
  override_csv = FALSE,
  database_connection = NULL
)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed tibble if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the
new parse operation}

\item{database_connection}{DBI connection object that holds a connection to
user defined database. If \code{save_table} is enabled without providing
value for this function an error will be thrown.}
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Identifier for the article being referenced.
  This is unique across all reference types (books, links, article).}
  \item{isbn}{ISBN identifying the textbook.}
  \item{citation}{A Textbook citation in a standard format.}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}
}
\description{
Return a list of text books that were used as references for drugs, carriers,
 enzymes, targets or transporters
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
drug_all()

# will throw an error, as database_connection is NULL
drug_all(save_table = TRUE)

# save in database in SQLite in memory database and return parsed tibble
sqlite_con <- DBI::dbConnect(RSQLite::SQLite())
drug_all(save_table = TRUE, database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in current location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_csv = TRUE)

# save in database, save parsed tibble as csv,
# if it does not exist in current location and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_table = TRUE, save_csv = TRUE,
database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in given location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in current location and
# return parsed tibble.
# if the csv exist override it and return it.
drug_all(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}
}
\seealso{
Other references: 
\code{\link{articles}},
\code{\link{links}}
}
\concept{references}
