% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_actions_parsers.R
\name{cett_actions_doc}
\alias{cett_actions_doc}
\alias{carriers_actions}
\alias{enzymes_actions}
\alias{targets_actions}
\alias{transporters_actions}
\title{Carriers/ Enzymes/ Targets/ Transporters Actions parsers}
\usage{
carriers_actions()

enzymes_actions()

targets_actions()

transporters_actions()
}
\value{
a tibble with 2 variables:
\describe{
  \item{action}{describe related action}
  \item{\emph{parent_id}}{carrier/ target/ enzyme/ transporter id}
}
}
\description{
Collection of related actions
}
\keyword{internal}
