% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_actions_parser.R
\name{cett_actions_doc}
\alias{cett_actions_doc}
\title{Carriers/ Enzymes/ Targets/ Transporters Actions parsers}
\value{
a tibble with 2 variables:
\describe{
  \item{action}{describe related action}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}
}
\description{
Collection of related actions
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
drug_all()

# will throw an error, as database_connection is NULL
drug_all(save_table = TRUE)

# save in database in SQLite in memory database and return parsed tibble
sqlite_con <- DBI::dbConnect(RSQLite::SQLite())
drug_all(save_table = TRUE, database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in current location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_csv = TRUE)

# save in database, save parsed tibble as csv,
# if it does not exist in current location and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_table = TRUE, save_csv = TRUE,
database_connection = sqlite_con)

# save parsed tibble as csv if it does not exist in given location,
# and return parsed tibble.
# if the csv exist before read it and return its data.
drug_all(save_csv = TRUE, csv_path = TRUE)

# save parsed tibble as csv if it does not exist in current location and
# return parsed tibble.
# if the csv exist override it and return it.
drug_all(save_csv = TRUE, csv_path = TRUE, override = TRUE)
}
}
\seealso{
Other cett: 
\code{\link{cett_doc}}
}
\concept{cett}
