% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_general_information_parsers.R
\name{cett_doc}
\alias{cett_doc}
\alias{carriers}
\alias{enzymes}
\alias{targets}
\alias{transporters}
\title{Carriers/ Enzymes/ Targets/ Transporters parsers}
\usage{
carriers()

enzymes()

targets()

transporters()
}
\value{
a tibble with 6 variables (8 for enzymes):
\describe{
  \item{id}{Universal Protein Resource (UniProt) Identifier for the record}
  \item{name}{related name}
  \item{organism}{Organism that the protein comes from.}
  \item{known_action}{Whether the pharmacological action of the drug is due
   to this target interaction.}
  \item{inhibition-strength}{Whether the strength of enzyme inhibition is
  strong, moderate, or unknown. \strong{Only applies to enzymes}}
  \item{induction-strength}{Whether the strength of enzyme induction is
  strong or unknown. \strong{Only applies to enzymes}}
  \item{position}{related position}
  \item{parent_id}{drugbank id}
}
}
\description{
Protein targets of drug action, enzymes that are inhibited/induced or
involved in metabolism, and carrier or transporter proteins involved in
movement of the drug across biological membranes.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_doc}},
\code{\link{cett_poly_pfms_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
