% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_others_parsers.R
\name{cett_ex_identity_doc}
\alias{cett_ex_identity_doc}
\alias{carriers_polypep_ex_ident}
\alias{enzymes_polypep_ex_ident}
\alias{targets_polypep_ex_ident}
\alias{transporters_polypep_ex_ident}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide External Identifiers
parsers}
\usage{
carriers_polypep_ex_ident()

enzymes_polypep_ex_ident()

targets_polypep_ex_ident()

transporters_polypep_ex_ident()
}
\value{
a tibble with 3 variables:
\describe{
  \item{resource}{Name of the source database.}
  \item{identifier}{Identifier for this drug in the given resource.}
  \item{parent_key}{polypeptide id}
}
}
\description{
Extract descriptions of identified polypeptide external identifiers for
targets, enzymes, carriers, or transporters.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_doc}},
\code{\link{cett_poly_pfms_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
