% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{cett_nodes_options}
\alias{cett_nodes_options}
\title{returns carriers, enzymes,targets and transporters node valid options.}
\usage{
cett_nodes_options()
}
\value{
list of ceett valid options
}
\description{
returns carriers, enzymes,targets and transporters node valid options.
}
\keyword{parsers}
