% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_general_infromation_parsers.R
\name{cett_poly_doc}
\alias{cett_poly_doc}
\alias{carriers_polypeptides}
\alias{enzymes_polypeptides}
\alias{targets_polypeptides}
\alias{transporters_polypeptides}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide parsers}
\usage{
carriers_polypeptides()

enzymes_polypeptides()

targets_polypeptides()

transporters_polypeptides()
}
\value{
a tibble with 20 variables:
\describe{
  \item{id}{\href{https://www.uniprot.org/}{Universal Protein Resource
  (UniProt) identifier}}
  \item{source}{Specifies whether the identified polypeptide ID is
  associated with any of the following UniProt knowledge bases:
  Swiss-Prot, which is manually annotated and reviewed, or TrEMBL,
  which is automatically annotated and not reviewed.}
  \item{name}{}
  \item{general_function}{General summary of the physiological function of
  the polypeptide}
  \item{specific_function}{A more specific description of the polypeptide’s
   physiological function within the cell.}
  \item{gene_name}{The short name commonly associated with the associated
   gene. Eg. PTGS1.}
  \item{locus}{The specific chromosomal location or position of the gene’s
   sequence on a chromosome.}
  \item{cellular_location}{The cellular location of the polypeptide.}
  \item{transmembrane_regions}{Areas of the polypeptide sequence that span
   a biological membrane.}
  \item{signal_regions}{Location of any signal peptides within the
   polypeptide sequence.}
  \item{theoretical_pi}{Theoretical isoelectric point.}
  \item{molecular_weight}{The molecular weight of the polypeptide.}
  \item{chromosome_location}{The chromosomal location of the polypeptide
  gene}
  \item{organism}{The organism in which this polypeptide functions.}
  \item{organism_ncbi_taxonomy_id}{}
  \item{amino_acid_sequence}{The amino acid sequence of the polypeptide}
  \item{amino_acid_format}{}
  \item{gene_sequence}{The sequence of the associated gene.}
  \item{gene_format}{}
  \item{parent_key}{carrier/ target/ enzyme/ transporter id}
}
}
\description{
Extract descriptions of identified polypeptide targets, enzymes, carriers,
 or transporters.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_pfms_doc}},
\code{\link{cett_poly_syn_doc}}
}
\concept{cett}
