% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cett_polypeptide_others_parsers.R
\name{cett_poly_syn_doc}
\alias{cett_poly_syn_doc}
\alias{carriers_polypeptides_syn}
\alias{enzymes_polypeptides_syn}
\alias{targets_polypeptides_syn}
\alias{transporters_polypeptides_syn}
\title{Carriers/ Enzymes/ Targets/ Transporters Polypeptide Synonyms parsers}
\usage{
carriers_polypeptides_syn()

enzymes_polypeptides_syn()

targets_polypeptides_syn()

transporters_polypeptides_syn()
}
\value{
a tibble with 2 variables:
\describe{
  \item{synonym}{}
  \item{parent_key}{polypeptide id}
}
}
\description{
Extract descriptions of identified polypeptide synonyms for targets,
 enzymes, carriers, or transporters.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other cett: 
\code{\link{cett_actions_doc}},
\code{\link{cett_doc}},
\code{\link{cett_ex_identity_doc}},
\code{\link{cett_go_doc}},
\code{\link{cett_poly_doc}},
\code{\link{cett_poly_pfms_doc}}
}
\concept{cett}
