% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbparser.R
\docType{package}
\name{dbparser}
\alias{dbparser}
\alias{dbparser-package}
\title{dbparser: A package for reading and parsing \strong{DrugBank} xml with the
option to save it in a given db.}
\description{
The main purpose of the `dbparser` package is to parse the
\strong{DrugBank} (\link{http://drugbank.ca}) database which is downloadable
in XML format from \link{https://www.drugbank.ca/releases/latest}.
The parsed data can then be explored and analyzed as desired by the user.
In addition to that user can save parsed dataframe into desired database as
well.
}
\details{
dbparser package provides three categories of important functions:
database related methods, xml db reader and \strong{DrugBank} elements
 parsers.
}
\section{database functions}{

 To open a connection to given database in order to store parsed
 \strong{DrugBank} elements database
}

\section{xml db reader functions}{

 Reads \strong{DrugBank} xml database and build drug elements full tree in
 database
}

\section{parsers functions}{

 Each parser function is responsible of parsing certain drug element and
 returning its dataframe
 with the ability to save it in a predefined database.
}

