% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_atc_codes_node_parser.R
\name{drug_atc_codes}
\alias{drug_atc_codes}
\title{Drug ATC Codes element parser}
\usage{
drug_atc_codes()
}
\value{
a tibble with 10 variables
}
\description{
The Anatomical Therapeutic Classification (ATC) code for the drug assigned
by the World Health Organization Anatomical Chemical Classification System.
}
\details{
Each `atc-code`` row has one or more level. The atc-code and level>
have a code  the code assigned by the World Health Organization Anatomical
Therapeutic Chemical Classification system.
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
