% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_calc_prop}
\alias{drug_calc_prop}
\title{Drug Calculated Properties parser}
\usage{
drug_calc_prop()
}
\value{
a tibble with 4 variables:
\describe{
 \item{kind}{Name of the property.}
 \item{value}{Predicted physicochemical properties; obtained by the use of
 prediction software such as ALGOPS and ChemAxon.}
 \item{source}{Name of the software used to calculate this property,
 either ChemAxon or ALOGPS.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Drug properties that have been predicted by ChemAxon or ALOGPS based on the
inputed chemical structure. Associated links below will redirect to
descriptions of the specific term.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
