% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_main_node_parser.R
\name{drug_general_information}
\alias{drug_general_information}
\title{Drugs General Information parser}
\usage{
drug_general_information()
}
\value{
a tibble with 15 variables:
\describe{
  \item{primary_key}{Drugbank id}
  \item{other_keys}{Other identifiers that may be associated with the drug}
  \item{type}{	Either small molecule, or biotech. Biotech is used for any
  drug that is derived from living systems or organisms, usually composed of
   high molecular weight mixtures of protein, while small molecule describes
    a low molecular weight organic compound.}
  \item{name}{}
  \item{created}{Date that this drug was first added to DrugBank.}
  \item{updated}{Denotes when this drug was last updated in DrugBank.}
  \item{description}{Descriptions of drug chemical properties,
   history and regulatory status.}
  \item{cas_number}{The Chemical Abstracts Service (CAS) registry number
   assigned to the drug.}
  \item{\emph{unii}}{Unique Ingredient Identifier (UNII) of this drug.}
  \item{average_mass}{The weighted average of the isotopic masses of the
  drug}
  \item{state}{One of solid, liquid, or gas}
  \item{monoisotopic_mass}{The mass of the most abundant isotope of the drug}
  \item{synthesis_reference}{Citation for synthesis of the drug molecule.}
  \item{fda_label}{Contains a URL for accessing the uploaded United States Food
  and Drug Administration (FDA) Monograph for this drug.}
  \item{msds}{Contains a URL for accessing the Material Safety Data Sheet
  (MSDS) for this drug.}
}
}
\description{
A description of the hierarchical chemical classification of the drug;
imported from ClassyFire.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
