% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_manufacturer_node_parser.R
\name{drug_manufacturers}
\alias{drug_manufacturers}
\title{Drug Manufacturers parser}
\usage{
drug_manufacturers()
}
\value{
a tibble with the following variables:
\describe{
 \item{generic}{A list of companies that are manufacturing the generic
  form of the drug.}
 \item{url}{A link to the companies that are manufacturing the drug.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
A list of companies that are manufacturing the commercially available forms
of this drug that are available in Canada and the Unites States.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
