% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{drug_node_options}
\alias{drug_node_options}
\title{returns durg node valid options.}
\usage{
drug_node_options()
}
\value{
list of \code{drug_element} valid options
}
\description{
returns durg node valid options.
}
\examples{
\dontrun{
drug_node_options()
}
}
\seealso{
Other common: 
\code{\link{references_node_options}()}
}
\concept{common}
\keyword{internal}
