% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{drug_node_options}
\alias{drug_node_options}
\title{returns drug node valid options.}
\usage{
drug_node_options()
}
\value{
list of drug valid options
}
\description{
returns drug node valid options.
}
\keyword{parsers}
