% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_parsers.R
\name{drug_patents}
\alias{drug_patents}
\title{Drug Patents parser
A property right issued by the United States Patent and Trademark
Office (USPTO) to an inventor for a limited time, in exchange for public
disclosure of the invention when the patent is granted. Drugs may be issued
multiple patents.}
\usage{
drug_patents()
}
\value{
a tibble with the following variables:
\describe{
 \item{number}{The patent number(s) associated with the drug.}
 \item{country}{The country that issued the patent rights.}
 \item{approved}{The date that the patent request was filed.}
 \item{expires}{The date that the patent rights expire.}
 \item{pediatric-extension}{Indicates whether or not a pediatric extension has been approved for
  the patent. Granted pediatric extensions provide an additional 6 months of
  market protection.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Drug Patents parser
A property right issued by the United States Patent and Trademark
Office (USPTO) to an inventor for a limited time, in exchange for public
disclosure of the invention when the patent is granted. Drugs may be issued
multiple patents.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_reactions}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
