% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_pathway_node_parsers.R
\name{drug_pathway_drugs}
\alias{drug_pathway_drugs}
\title{Drug Pathway Drugs parser}
\usage{
drug_pathway_drugs()
}
\value{
a tibble with pathway drugsproperties
}
\description{
Drugs involved in this pathway.
}
\section{read_drugbank_xml_db}{

\code{\link{read_drugbank_xml_db}} function must be called first before any
parser.

If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}

\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other pathway: 
\code{\link{drug_pathway_enzyme}()},
\code{\link{drug_pathway}()}
}
\concept{pathway}
