% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_reaction_node_parser.R
\name{drug_reactions}
\alias{drug_reactions}
\title{Drug Reactions Parsers}
\usage{
drug_reactions()
}
\value{
a tibble with 5 variables:
\describe{
  \item{sequence}{	Reactions are displayed within a numerical sequence}
  \item{left_drugbank_name}{The substrate of the reaction. Maybe a drug or a
   metabolite.}
  \item{rightt_drugbank_name}{	The product of the reaction. Maybe a drug or a
   metabolite.}
  \item{left_drugbank_id}{}
  \item{right_drugbank_id}{}
  \item{parent_id}{drugbank id}
}
}
\description{
Extract the sequential representation of the metabolic reactions that this
 drug molecule is involved in. Depending on available information, this may
 include metabolizing enzymes, reaction type, substrates, products,
 pharmacological activity of metabolites, and a structural representation of
  the biochemical reactions.
}
\seealso{
Other drugs: 
\code{\link{drug_affected_organisms}()},
\code{\link{drug_ahfs_codes}()},
\code{\link{drug_atc_codes}()},
\code{\link{drug_calc_prop}()},
\code{\link{drug_categories}()},
\code{\link{drug_classification}()},
\code{\link{drug_dosages}()},
\code{\link{drug_ex_identity}()},
\code{\link{drug_exp_prop}()},
\code{\link{drug_external_links}()},
\code{\link{drug_food_interactions}()},
\code{\link{drug_general_information}()},
\code{\link{drug_groups}()},
\code{\link{drug_interactions}()},
\code{\link{drug_intern_brand}()},
\code{\link{drug_manufacturers}()},
\code{\link{drug_mixtures}()},
\code{\link{drug_packagers}()},
\code{\link{drug_patents}()},
\code{\link{drug_pdb_entries}()},
\code{\link{drug_pharmacology}()},
\code{\link{drug_prices}()},
\code{\link{drug_products}()},
\code{\link{drug_reactions_enzymes}()},
\code{\link{drug_salts}()},
\code{\link{drug_sequences}()},
\code{\link{drug_snp_adverse_reactions}()},
\code{\link{drug_snp_effects}()},
\code{\link{drug_syn}()}
}
\concept{drugs}
