% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_synonym_node_parser.R
\name{drug_syn}
\alias{drug_syn}
\title{Drug Synonyms parser}
\usage{
drug_syn()
}
\value{
a tibble with 3 variables:
\describe{
 \item{language}{Names of the drug in languages other than English.}
 \item{coder}{Organization or source providing the synonym. For example,
  INN indicates the synonym is an International Nonproprietary Name,
  while IUPAC indicates the synonym is the nomenclature designated by the
  International Union of Pure and Applied Chemistry.}
 \item{\emph{drugbank_id}}{drugbank id}
}
}
\description{
Other names or identifiers that are associated with this drug.
}
\keyword{internal}
