% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{drugs}
\alias{drugs}
\title{Run all drug  related parsers}
\usage{
drugs()
}
\value{
a list of all drugs parsed tibbles
}
\description{
Run all parsers that retrieve drugs related information
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other collective_parsers: 
\code{\link{cett}()},
\code{\link{run_all_parsers}()}
}
\concept{collective_parsers}
