% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references_parsers.R
\name{links}
\alias{links}
\alias{drugs_links}
\alias{carriers_links}
\alias{enzymes_links}
\alias{targets_links}
\alias{transporters_links}
\title{Drugs/ Carriers/ Enzymes/ Targets/ Transporters links element parser}
\usage{
drugs_links()

carriers_links()

enzymes_links()

targets_links()

transporters_links()
}
\value{
a tibble with 4 variables:
\describe{
  \item{ref-id}{Name of the source website}
  \item{title}{Identifier for this drug in the given resource}
  \item{url}{The url of the website}
  \item{\emph{parent_id}}{drug/ carrier/ target/ enzyme/ transporter id}
}
}
\description{
Return a list of websites that were used as references for
Drugs/ Carriers/ Enzymes/ Targets/ Transporters
}
\keyword{internal}
