% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_helpers.R
\name{merge_drugbank_twosides}
\alias{merge_drugbank_twosides}
\title{Merge a DrugBank dvobject with a TWOSIDES dvobject}
\usage{
merge_drugbank_twosides(db_object, twosides_db)
}
\arguments{
\item{db_object}{A dvobject from `parseDrugBank()` or an existing merged dvobject.}

\item{twosides_db}{A dvobject from `parseTWOSIDES()`.}
}
\value{
A new, nested dvobject with the TWOSIDES data added.
}
\description{
Integrates drug-drug interaction data from TWOSIDES with the rich mechanistic
information from DrugBank. This function is chainable and can accept a raw
DrugBank object or an already-merged dvobject.
}
\seealso{
Other mergers: 
\code{\link{merge_drugbank_onsides}()}
}
\concept{mergers}
