% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{parse_drug_element}
\alias{parse_drug_element}
\title{Extracts the given drug elements and return data as list of dataframes.}
\usage{
parse_drug_element(elements_options = c("all"), save_table = FALSE,
  save_csv = FALSE, csv_path = ".", override_csv = FALSE)
}
\arguments{
\item{elements_options}{list,  options of elements to be parsed. Default is "all"}

\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed dataframe if true}

\item{csv_path}{location to save csv files into it, default is current location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in the new parse operation}
}
\value{
list of selected drug elements dataframes
}
\description{
\code{parse_drug_element} returns list of dataframes of drugs selected elements.
}
\details{
This functions extracts selected element of drug nodes in drug bank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{get_xml_db_rows}} function like
any other parser function.
If \code{\link{get_xml_db_rows}} is called before for any reason, so
no need to call it again before calling this function.

parse_drug_element_options can be called to know the valid options for
this method
}
\examples{
\donttest{
# return only the parsed dataframe
parse_drug_element()

# save in database and return parsed dataframe
parse_drug_element(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current location and
# return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_element(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not
# exist in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_element(ssave_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given location and
# return parsed dataframe.
# If the csv exist before read it and return its data.
parse_drug_element(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current
# location and return parsed dataframe.
# If the csv exist override it and return it.
parse_drug_element(save_csv = TRUE, csv_path = TRUE, override = TRUE)
parse_drug_element(c("drug_ahfs_codes", "drug_carriers"), save_table = TRUE)
parse_drug_element(save_table = FALSE)
parse_drug_element(c("drug_ahfs_codes", "drug_carriers"))
}
}
