% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{parseDrugBank}
\alias{parseDrugBank}
\title{parseDrugBank
 parses given DrugBank XML database into a dvobject. dvobject is a list of
 data.frames in which each data.frame represents a part of parsed data (i.e
 drugs, prices, carriers, ...)}
\usage{
parseDrugBank(
  db_path,
  drug_options = drug_node_options(),
  parse_salts = TRUE,
  parse_products = TRUE,
  references_options = references_node_options(),
  cett_options = cett_nodes_options()
)
}
\arguments{
\item{db_path}{\strong{string}, full path for the \strong{DrugBank} xml or
zip file.}

\item{drug_options}{\strong{character vector}, list of sub drug related nodes
names options to parse (default = all available nodes)}

\item{parse_salts}{\strong{boolean}, parse salts info (default = TRUE)}

\item{parse_products}{\strong{boolean}, parse products info (default = TRUE)}

\item{references_options}{\strong{character vector}, list of sub references
related nodes names options to parse (default = all available nodes)}

\item{cett_options}{\strong{character vector}, list of sub cett related nodes
names options to parse (default = all available nodes)}
}
\value{
dvobject
}
\description{
parseDrugBank
 parses given DrugBank XML database into a dvobject. dvobject is a list of
 data.frames in which each data.frame represents a part of parsed data (i.e
 drugs, prices, carriers, ...)
}
\concept{parsers}
