% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{parseDrugBank}
\alias{parseDrugBank}
\title{parseDrugBank
 parses given DrugBank XML database into a dvobject. dvobject is a list of
 data.frames in which each data.frame represents a part of parsed data (i.e
 drugs, prices, carriers, ...)}
\usage{
parseDrugBank(
  db_path,
  drug_options = drug_node_options(),
  parse_salts = TRUE,
  parse_products = TRUE,
  references_options = references_node_options(),
  cett_options = c("carriers", "enzymes", "targets", "transporters")
)
}
\arguments{
\item{db_path}{\strong{string}, full path for the
\strong{DrugBank} xml or zip file.}
}
\value{
dvobject
}
\description{
parseDrugBank
 parses given DrugBank XML database into a dvobject. dvobject is a list of
 data.frames in which each data.frame represents a part of parsed data (i.e
 drugs, prices, carriers, ...)
}
\concept{parsers}
