% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_common_utilities.R
\name{read_drugbank_xml_db}
\alias{read_drugbank_xml_db}
\title{Reads \strong{DrugBank} xml database and load it in memory.}
\usage{
read_drugbank_xml_db(drugbank_db_path)
}
\arguments{
\item{drugbank_db_path}{string, full path for the \strong{DrugBank} xml or
zip file.}
}
\value{
TRUE when sets the db tree in memory to be used by parser
methods and FALSE otherwise
}
\description{
\code{read_drugbank_xml_db} sets \strong{DrugBank} db xml full tree in memory
}
\details{
This functions reads \strong{DrugBank} xml database and sets the full tree
 save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
in memory directly without returning it.
It must be called one before using parser functions, and once it is called
If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again.
}
\examples{
\dontrun{
read_drugbank_xml_db("db_full_path")
read_drugbank_xml_db(drugbank_db_path = "db_full_path")
}
}
