% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_common_utilities.R
\name{read_drugbank_xml_db}
\alias{read_drugbank_xml_db}
\title{Reads \strong{DrugBank} xml database and load it into memory.}
\usage{
read_drugbank_xml_db(db_path)
}
\arguments{
\item{db_path}{\strong{string}, full path for the
\strong{DrugBank} xml or zip file.}
}
\value{
loaded DB or NULL
}
\description{
\code{read_drugbank_xml_db} loads \strong{DrugBank} xml database full tree
into memory.
}
\details{
This functions reads \strong{DrugBank} xml database and load it into memory
 for later processing. Hence; this method \strong{must} be called before any
 other function in the package and it needs to be called one time only.

It takes one single mandatory argument which is the location of DrugBank db.
}
\keyword{internal}
