% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugbank_parser.R
\name{references_node_options}
\alias{references_node_options}
\title{returns references node valid options.}
\usage{
references_node_options()
}
\value{
list of \code{drug_element} valid options
}
\description{
returns references node valid options.
}
\keyword{internal}
