% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{references_node_options}
\alias{references_node_options}
\title{returns references node valid options.}
\usage{
references_node_options()
}
\value{
list of \code{drug_element} valid options
}
\description{
returns references node valid options.
}
\examples{
\dontrun{
references_node_options()
}
}
\seealso{
Other common: 
\code{\link{drug_element_options}()},
\code{\link{drug_element}()},
\code{\link{drug_node_options}()},
\code{\link{run_all_parsers}()}
}
\concept{common}
\keyword{internal}
