% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_all_nodes_parser.R
\name{run_all_parsers}
\alias{run_all_parsers}
\title{extracts the all drug elements and return data as list of tibbles.}
\usage{
run_all_parsers()
}
\value{
all drug elements tibbles
}
\description{
this functions extracts all element of drug nodes in \strong{DrugBank}
xml database.
}
\examples{
\dontrun{
# the same parameters and usage will be applied for any parser
# return only the parsed tibble
run_all_parsers()

}
}
\seealso{
Other common: 
\code{\link{drug_element_options}()},
\code{\link{drug_element}()},
\code{\link{drug_node_options}()},
\code{\link{references_node_options}()}

Other collective_parsers: 
\code{\link{parse_cett_node}()},
\code{\link{parse_drug_nodes}()}
}
\concept{collective_parsers}
\concept{common}
