% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvobject_metadata.R
\name{show_dvobject_metadata}
\alias{show_dvobject_metadata}
\title{init_dvobject
  Returns data.frame with two columns (key, value) of dvobject attributes}
\usage{
show_dvobject_metadata(dvobject)
}
\arguments{
\item{dvobject}{- dvobject list to show related metadata}
}
\value{
data.frame
}
\description{
init_dvobject
  Returns data.frame with two columns (key, value) of dvobject attributes
}
\concept{utility}
