% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_transporter_node_parser.R
\name{transporters_actions}
\alias{transporters_actions}
\title{Extracts the drug transporters actions element and return data as data frame.}
\usage{
transporters_actions(save_table = FALSE, save_csv = FALSE,
  csv_path = ".", override_csv = FALSE)
}
\arguments{
\item{save_table}{boolean, save table in database if true.}

\item{save_csv}{boolean, save csv version of parsed dataframe if true}

\item{csv_path}{location to save csv files into it, default is current
location, save_csv must be true}

\item{override_csv}{override existing csv, if any, in case it is true in
the new parse operation}
}
\value{
drug transporters actions node attributes date frame
}
\description{
\code{transporters actions} returns data frame of drug
 transporters
actions elements.
}
\details{
This functions extracts the transporters actions element of drug node in
drugbank
xml database with the option to save it in a predefined database via
\code{\link{open_db}} method. It takes one single optional argument to
save the returned dataframe in the database.
It must be called after \code{\link{read_drugbank_xml_db}} function like
any other parser function.
If \code{\link{read_drugbank_xml_db}} is called before for any reason, so
no need to call it again before calling this function.
}
\examples{
\dontrun{
# return only the parsed dataframe
transporters_actions()

# save in database and return parsed dataframe
transporters_actions(save_table = TRUE)

# save parsed dataframe as csv if it does not exist in current location
# and return parsed dataframe.
# If the csv exist before read it and return its data.
transporters_actions(save_csv = TRUE)

# save in database, save parsed dataframe as csv if it does not exist
# in current location and return parsed dataframe.
# If the csv exist before read it and return its data.
transporters_actions(save_table = TRUE, save_csv = TRUE)

# save parsed dataframe as csv if it does not exist in given location
# and return parsed dataframe.
# If the csv exist before read it and return its data.
transporters_actions(save_csv = TRUE, csv_path = TRUE)

# save parsed dataframe as csv if it does not exist in current location
#  and return parsed dataframe.
# If the csv exist override it and return it.
transporters_actions(
  save_csv = TRUE, csv_path = TRUE,
  override = TRUE
)
}
}
\seealso{
Other transporters: \code{\link{transporters_articles}},
  \code{\link{transporters_links}},
  \code{\link{transporters_polypep_ex_ident}},
  \code{\link{transporters_polypeptide_go}},
  \code{\link{transporters_polypeptide_pfams}},
  \code{\link{transporters_polypeptide_syn}},
  \code{\link{transporters_polypeptide}},
  \code{\link{transporters_textbooks}},
  \code{\link{transporters}}
}
\concept{transporters}
